+++ test/countries/atlas_engine/pl/country_profile_test.rb
@@ -10,6 +10,15 @@ module AtlasEngine
         @profile = CountryProfile.for("PL")
       end
 
+      test "#correctors value is correct for source open-address" do
+        assert_equal [
+          Pl::AddressImporter::Corrections::OpenAddress::CityCorrector,
+          Pl::AddressImporter::Corrections::OpenAddress::PostalCodePlaceholderCorrector,
+          Pl::AddressImporter::Corrections::OpenAddress::EmptyStreetCorrector,
+        ],
+          @profile.ingestion.correctors(source: "open_address")
+      end
+
       test "#address_parser returns the custom PL parser" do
         assert_equal Pl::ValidationTranscriber::AddressParser,
           @profile.validation.address_parser